# swagger/swagger.yaml
swagger: '2.0'
info:
  description: Periscol Backend API documentation
  version: 0.0.1
  title: Periscol Backend API
  contact: {}
  license:
    name: unlicensed
host: localhost:8080
basePath: /
...
paths:
...
  /api/families:
    get:
      tags:
        - family-resource
      summary: getAllFamilies
      operationId: getAllFamiliesUsingGET
      produces:
        - '*/*'
      responses:
        '200':
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/FamilyDTO'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
  /api/family:
    post:
      tags:
        - family-resource
      summary: createFamily
      operationId: createFamilyUsingPOST
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - in: body
          name: familyDTO
          description: familyDTO
          required: true
          schema:
            $ref: '#/definitions/FamilyDTO'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/FamilyDTO'
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
  /api/family/{id}:
    get:
      tags:
        - family-resource
      summary: getFamily
      operationId: getFamilyUsingGET
      produces:
        - '*/*'
      parameters:
        - name: id
          in: path
          description: id
          required: true
          type: integer
          format: int64
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/FamilyDTO'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
    put:
      tags:
        - family-resource
      summary: updateFamily
      operationId: updateFamilyUsingPUT
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - in: body
          name: familyDTO
          description: familyDTO
          required: true
          schema:
            $ref: '#/definitions/FamilyDTO'
        - name: id
          in: path
          description: id
          required: true
          type: integer
          format: int64
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/FamilyDTO'
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
    delete:
      tags:
        - family-resource
      summary: deleteFamily
      operationId: deleteFamilyUsingDELETE
      produces:
        - '*/*'
      parameters:
        - name: id
          in: path
          description: id
          required: true
          type: integer
          format: int64
      responses:
        '200':
          description: OK
        '204':
          description: No Content
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
    patch:
      tags:
        - family-resource
      summary: partialUpdateFamily
      operationId: partialUpdateFamilyUsingPATCH
      consumes:
        - application/json
        - application/merge-patch+json
      produces:
        - '*/*'
      parameters:
        - in: body
          name: familyDTO
          description: familyDTO
          required: true
          schema:
            $ref: '#/definitions/FamilyDTO'
        - name: id
          in: path
          description: id
          required: true
          type: integer
          format: int64
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/FamilyDTO'
        '204':
          description: No Content
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
...
definitions:
...
  FamilyDTO:
    type: object
    properties:
      id:
        type: integer
        format: int64
      postalAdress:
        type: string
      referingParentName:
        type: string
      referingParentSurname:
        type: string
      telephoneNumber:
        type: string
    title: FamilyDTO
...